Add UniswapV2Pair contract
The UniswapV2Pair contract is a core contract that tracks the reserves
of two tokens and facilitate swaps based on reserve balances.

Liquidity is seeded to the Pairs by sending amounts of the tokens it tracks
in such a way that it determines the swap price from a token to another.
Lastly, Pairs track liquidity providers shares in the pool of token reserves
by minting Liquidity Pool tokens when a user adds liquidity. It subsequently
burns those tokens (LP Tokens) when a user wishes to redeem the tokens deposited.